Moved test classes into their own package so that the tests package can be run as a Unit test to run all of the individual classes as a test suite.
