+++ src/test/java/net/sigmalab/maven/plugin/jira/testClasses/ReleaseVersionMojoTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package net.sigmalab.maven.plugin.jira.tests;
+package net.sigmalab.maven.plugin.jira.testClasses;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
